+++ src/core/actions/look.cpp
@@ -56,11 +56,13 @@ namespace trogdor {
 
          else {
 
-            entity::ThingList items;
+            std::list<entity::Thing *> items;
 
             // consider matching inventory items, if there are any
-            for (auto const &invObject: player->getInventoryObjectsByName(object)) {
-                  items.push_front(invObject);
+            for (auto const &invObjectPtr: player->getInventoryObjectsByName(object)) {
+                  if (auto const &invObject = invObjectPtr.lock()) {
+                     items.push_front(invObject.get());
+                  }
             };
 
             // also consider matching items in the room, if there are any